Add version for dotnet core
Refactor local adapter
