+++ src/XmlFormatter/Update/LocalVersionReciever.cs
@@ -1,41 +1,20 @@
-﻿using PluginFramework.EventMessages;
-using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Reflection;
-using System.Threading.Tasks;
-using XmlFormatterModel.Update;
-using XmlFormatterModel.Update.Strategies;
+﻿using System.Reflection;
+using XmlFormatterModel.Update.Adapter;
 
 namespace XmlFormatter.Update
 {
-    class LocalVersionReciever : IVersionRecieverStrategy
+    class LocalVersionReciever : LocalResourceVersionReciever
     {
-        public event EventHandler<BaseEventArgs> Error;
-
-        public Task<IRelease> GetLatestRelease()
+        /// <inheritdoc/>
+        protected override Assembly GetAssembly()
         {
-            return default;
+            return Assembly.GetExecutingAssembly();
         }
 
-        public Task<List<IRelease>> GetReleases()
+        /// <inheritdoc/>
+        protected override string GetResourcePath()
         {
-            return default;
-        }
-
-        public async Task<Version> GetVersion(IVersionConvertStrategy convertStrategy)
-        {
-            Assembly assembly = Assembly.GetExecutingAssembly();
-            string versionString = "0.0.0";
-            using (Stream stream = assembly.GetManifestResourceStream("XmlFormatter.Version.txt"))
-            {
-                using (StreamReader reader = new StreamReader(stream))
-                {
-                    versionString = reader.ReadLine();
-                }
-            }
-
-            return convertStrategy.ConvertStringToVersion(versionString);
+            return "XmlFormatter.Version.txt";
         }
     }
 }
