+++ src/XmlFormatterOsIndependent/XmlFormatterOsIndependent.csproj
@@ -27,12 +27,16 @@
   <ItemGroup>
     <None Remove="Assets\HotfolderIcon.ico" />
     <None Remove="Assets\XmlFormatterSettingsIcon.ico" />
+    <None Remove="version.txt" />
     <None Remove="XmlFormatterIcon.ico" />
   </ItemGroup>
   <ItemGroup>
     <Content Include="Assets\HotfolderIcon.ico" />
     <Content Include="Assets\XmlFormatterSettingsIcon.ico" />
     <Content Include="XmlFormatterIcon.ico" />
+  </ItemGroup>
+  <ItemGroup>
+    <EmbeddedResource Include="Version.txt" />
   </ItemGroup>
   <ItemGroup>
     <PackageReference Include="Avalonia" Version="0.9.10" />
